+++ Pipfile
@@ -1,0 +1,51 @@
+[[source]]
+name = "pypi"
+url = "https://pypi.org/simple"
+verify_ssl = true
+
+[dev-packages]
+autopep8 = "*"
+
+[packages]
+blinker = "==1.4"
+cachelib = "==0.1.1"
+certifi = "==2020.6.20"
+chardet = "==3.0.4"
+click = "==7.1.2"
+dominate = "==2.5.1"
+flask-session = "==0.3.2"
+gunicorn = "==20.0.4"
+heroku = "==0.1.4"
+idna = "==2.10"
+itsdangerous = "==1.1.0"
+passlib = "==1.7.2"
+postgres = "==3.0.0"
+psycopg2 = "==2.8.5"
+psycopg2-binary = "==2.8.5"
+psycopg2-pool = "==1.1"
+python-dateutil = "==1.5"
+python-dotenv = "==0.13.0"
+pytz = "==2020.1"
+requests = "==2.24.0"
+speaklater = "==1.3"
+urllib3 = "==1.25.9"
+visitor = "==0.1.3"
+Babel = "==2.8.0"
+Flask = "==1.1.2"
+Flask-BabelEx = "==0.9.4"
+Flask-Bootstrap = "==3.3.7.1"
+Flask-DotEnv = "==0.1.2"
+Flask-Login = "==0.5.0"
+Flask-Mail = "==0.9.1"
+Flask-Principal = "==0.4.0"
+Flask-Security = "==3.0.0"
+Flask-SQLAlchemy = "==2.4.3"
+Flask-WTF = "==0.14.3"
+Jinja2 = "==2.11.2"
+MarkupSafe = "==1.1.1"
+SQLAlchemy = "==1.3.17"
+Werkzeug = "==1.0.1"
+WTForms = "==2.3.1"
+
+[requires]
+python_version = "3.8"
