--- application/templates/delete.html
@@ -1,119 +1,0 @@
-{% extends "base.html" %}
-{% block title %}Delete Patient{% endblock%}
-
-{% block head%}
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" href="https://unpkg.com/purecss@2.0.3/build/pure-min.css" integrity="sha384-cg6SkqEOCV1NbJoCu11+bm0NvBRc8IYLRGXkmNrqUBfTjmMYwNKPWBTIKyw9mHNJ" crossorigin="anonymous">
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
-    <script src="{{url_for('static', filename='state.js')}}"></script>
-  <style>
-      .header {
-    padding: 4px;
-    text-align: center;
-    background: #1abc9c;
-    color: white;
-    font-size: 20px;
-    position: absoulute;
-    }
-
-    .pure-menu-link{
-        background: rgb(66, 184, 221);
-        
-        text-shadow: 0 1px 1px rgba(0, 0, 0, 0.2);
-    }
-    .pure-menu-list{
-        padding-bottom: 10px;
-    }
-    .search-container button:hover {
-    background: #ccc;
-  }
-  .search-container {
-      float: none;
-    }
-  </style>
-{% endblock %}
-{% block body %}
-    <body>
-      <h1 style="text-align: center;">Delete Patient</h1>
-      <form style="padding-left: 380px" class="pure-form pure-form-aligned" action="{% if task=='search' %}{{ url_for('getPatient',  feature='delete') }} {% elif task=='delete' %}{{ url_for('deletePatient') }} {% endif %}" method='post'>
-
-    <fieldset>
-        <div class="pure-control-group">
-            <label for="aligned-SSN">Patient ID *</label>
-            {% if task=='search' %}
-                <input type="text" id="SSN" name="search" placeholder="" />
-                <button type="get" name="get" class="pure-button pure-button-primary">Get</button>
-            {% elif task=='delete' %}
-                <input type="text" id="SSN" placeholder="" name="search" value="{{ patientid }}" readonly/>
-            {% endif %}
-        </div>
-        <div class="pure-control-group">
-            <label for="aligned-name">Patient Name*</label>
-            <input type="text" name="name" id="name" value="{{ name }}" />
-        </div>
-        <div class="pure-control-group">
-            <label for="aligned-age">Patient Age *</label>
-            <input type="number" id="age" name="age" value="{{ age }}" />
-        </div>
-        <div class="pure-control-group">
-            <label for="aligned-admission">Date of Admission *</label>
-            <input type="date" name="dateOfAdmission" id="admission" value="{{ dateOfAdmission }}"/>
-        </div>
-        <div><label style="padding-left: 80px" for="aligned-states">Type of bed *</label>
-         <select name="typeOfBed">
-            <option value="">Select type..</option>
-            {% if typeOfBed  == 'type-1' %}
-                <option value="type-1" selected>type-1</option>
-                <option value="type-2">type-2</option>
-                <option value="type-3">type-3</option>
-            {% elif typeOfBed  == 'type-2' %}
-                <option value="type-1">type-1</option>
-                <option value="type-2" selected>type-2</option>
-                <option value="type-3">type-3</option>
-            {% elif typeOfBed  == 'type-3' %}
-                <option value="type-1">type-1</option>
-                <option value="type-2">type-2</option>
-                <option value="type-3" selected>type-3</option>
-            {% endif %}
-         </select>
-     </div><br>
-              <div class="pure-control-group">
-            <label for="aligned-address">Address *</label>
-            <input type="textarea" name="address" value="{{ address }}"></input>
-            </div>
-
-          <div class="pure-control-group">
-      <label for="State">State</label>
-        <select id="select1"  name="state" onchange="populate(this.id,'select2')">
-            <option>Select</option>
-            <option value="Delhi" {% if state == 'Delhi' %}selected{% endif%}> Delhi</option>
-            <option value="Haryana" {% if state == 'Haryana' %}selected{% endif%}>Haryana</option>
-            <option value="Himachal Pradesh" {% if state == 'Himachal Pradesh' %}selected{% endif%}>Himachal Pradesh</option>
-            <option value="Jammu and Kashmir" {% if state == 'Jammu and Kashmir' %}selected{% endif%}>Jammu and Kashmir</option>
-            <option value="Ladakh" {% if state == 'Ladakh' %}selected{% endif%}>Ladakh</option>
-            <option value="Punjab" {% if state == 'Punjab' %}selected{% endif%}>Punjab</option>
-            <option value="Rajasthan" {% if state == 'Rajasthan' %}selected{% endif%}>Rajasthan</option>
-            <option value="Uttarakhand" {% if state == 'Uttarakhand' %}selected{% endif%}>Uttarakhand</option>
-            <option value="Uttarpradesh" {% if state == 'Uttarpradesh' %}selected{% endif%}>Uttar Pradesh</option>
-        </select>
-    </div>     <div style="padding-left: 150px">
-                <td>City</td>
-                    <td>
-                        <select  id="select2" name="city">
-                            <option value="">Select</option>
-
-                        </select>
-                    </td>
-                </div>
-             <div class="pure-controls">
-            {% if task=="delete" %}
-            <button type="Delete" class="pure-button pure-button-primary">Delete</button>
-            {% endif %}
-        </div>
-    </fieldset>
-</form>
-</body>
-</html>
-
-    </div>
-{% endblock %}
