--- application/templates/index.html
@@ -1,62 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-	<meta name="viewport" content="width=device-width, initial-scale=1">
-	<link rel="stylesheet" href="https://unpkg.com/purecss@2.0.3/build/pure-min.css" integrity="sha384-cg6SkqEOCV1NbJoCu11+bm0NvBRc8IYLRGXkmNrqUBfTjmMYwNKPWBTIKyw9mHNJ" crossorigin="anonymous">
-    <!-- Latest compiled and minified CSS -->
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.5.0/css/bootstrap.min.css">
-	<title>
-		Log In
-	</title>
-	
-<style>
-	.header {
-  padding: 10px;
-  text-align: center;
-  background: #1abc9c;
-  color: white;
-  font-size: 20px;
-  position: absoulute;
-}
-
-.pure-form{
-	padding-left: 400px;
-	padding-top: 80px;
-}
-
-
-</style>
-</head>
-
-
-<body>
-	<div class="header">
-    <h4>ABC Hospital Management System</h4>
-    </div>
-    {% with messages = get_flashed_messages(with_categories=true) %}
-        {% if messages %}
-            {% for category, message in messages %}
-                <div class = "alert alert-{{category}}">
-                  <span>{{message}}</span>
-                </div>
-            {% endfor %}
-        {% endif %}
-    {% endwith %}
-    <form class="pure-form pure-form-aligned" action="" method="post" >
-    <fieldset>
-        <div class="pure-control-group">
-            <label for="name">UserID</label>
-            <input type="text" name="userid" placeholder="Userid" />
-        </div>
-        <div class="pure-control-group">
-            <label for="password">Password</label>
-            <input type="password" name="password"  placeholder="Password" />
-        </div>
-        <div class="pure-controls">
-            <button type="submit" class="pure-button pure-button-primary">Submit</button>
-        </div>
-      
-    </fieldset>
-</form>
-</body>
-</html>
