--- application/templates/view.html
@@ -1,78 +1,0 @@
-{% extends "base.html" %}
-{% block title %}View Patient{% endblock%}
-
-{% block head%}
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <link rel="stylesheet" href="https://unpkg.com/purecss@2.0.3/build/pure-min.css" integrity="sha384-cg6SkqEOCV1NbJoCu11+bm0NvBRc8IYLRGXkmNrqUBfTjmMYwNKPWBTIKyw9mHNJ" crossorigin="anonymous">
-  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
-  <style>
-      .header {
-    padding: 4px;
-    text-align: center;
-    background: #1abc9c;
-    color: white;
-    font-size: 20px;
-    position: absoulute;
-    }
-
-    .pure-menu-link{
-        background: rgb(66, 184, 221);
-        
-        text-shadow: 0 1px 1px rgba(0, 0, 0, 0.2);
-    }
-    .pure-menu-list{
-        padding-bottom: 10px;
-    }
-    .search-container button:hover {
-    background: #ccc;
-  }
-  .search-container {
-      float: none;
-    }
-  </style>
-{% endblock %}
-{% block body %}
-    <body>
-      <h1 style="text-align: center;">View Patient</h1>
-      <form style="padding-left: 380px" class="pure-form pure-form-aligned" action='{{ url_for('viewPatient') }}' method="post">
-
-    <fieldset>
-        <div class="pure-control-group">
-            <label for="aligned-SSN">Patient ID *</label>
-            <input type="text" value="{{ patientid }}" name="search">
-      <button type="submit"><i class="fa fa-search"></i></button>
-        </div>
-        <div class="pure-control-group">
-            <label for="aligned-name">Patient Name*</label>
-            <input type="text" id="name" value="{{ name }}" readonly/>
-        </div>
-        <div class="pure-control-group">
-            <label for="aligned-age">Patient Age *</label>
-            <input type="number" id="age" value="{{ age }}" readonly/>
-        </div>
-        <div class="pure-control-group">
-            <label for="aligned-admission">Date of Admission *</label>
-            <input type="text" id="admission" value="{{ dateOfAdmission }}"  readonly/>
-        </div>
-        <div><label style="padding-left: 80px" for="aligned-states">Type of bed *</label>
-            <input type="text" id="bed" value="{{ typeOfBed }}" readonly/>
-     </div><br>
-              <div class="pure-control-group">
-            <label for="aligned-address">address *</label>
-            <input type="text" value='{{ address }}' readonly></input>
-            </div>
-
-          <div class="pure-control-group">
-        <label for="State">State</label>
-        <input type="text" id="bed" value="{{ state }}" readonly />
-    </div>     
-    <div style="padding-left: 150px">
-      <td>City</td>
-      <td>
-        <input type="text" id="bed" value="{{ city }}" readonly/>
-      </td>
-    </div>
-    </fieldset>
-</form>
-    </div>
-{% endblock %}
