--- main.py
@@ -1,1 +1,0 @@
-from application import app
