+++ requirements.txt
@@ -1,3 +1,4 @@
+autopep8==1.5.4
 Babel==2.8.0
 blinker==1.4
 cachelib==0.1.1
@@ -27,12 +28,14 @@ postgres==3.0.0
 psycopg2==2.8.5
 psycopg2-binary==2.8.5
 psycopg2-pool==1.1
+pycodestyle==2.6.0
 python-dateutil==1.5
 python-dotenv==0.13.0
 pytz==2020.1
 requests==2.24.0
 speaklater==1.3
 SQLAlchemy==1.3.17
+toml==0.10.1
 urllib3==1.25.9
 visitor==0.1.3
 Werkzeug==1.0.1
