+++ templates/add_diagnostics.html
