+++ templates/add_medicine.html
