+++ templates/create_patient.html
@@ -1,0 +1,94 @@
+{% extends "base.html" %}
+{% block title %} Create Patient{% endblock%}
+{% block head%}
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <link rel="stylesheet" href="https://unpkg.com/purecss@2.0.3/build/pure-min.css" integrity="sha384-cg6SkqEOCV1NbJoCu11+bm0NvBRc8IYLRGXkmNrqUBfTjmMYwNKPWBTIKyw9mHNJ" crossorigin="anonymous">
+    <script src="{{url_for('static', filename='state.js')}}"></script>
+    <style>
+        .header {
+    padding: 4px;
+    text-align: center;
+    background: #1abc9c;
+    color: white;
+    font-size: 20px;
+    position: absoulute;
+    }
+
+    .pure-menu-link{
+        background: rgb(66, 184, 221);
+        
+        text-shadow: 0 1px 1px rgba(0, 0, 0, 0.2);
+    }
+    .pure-menu-list{
+        padding-bottom: 10px;
+    }
+
+    </style>
+{% endblock %}
+
+{% block body %}
+<body>
+	<form style="padding-left: 380px" class="pure-form pure-form-aligned" method="post" action="{{ url_for('add_patient') }}">
+    <fieldset>
+        <div class="pure-control-group">
+            <label for="aligned-SSN">Patient SSN ID *</label>
+            <input type="text" id="SSN" placeholder="" name="ssnId"/>
+        </div>
+        <div class="pure-control-group">
+            <label for="aligned-name">Patient Name*</label>
+            <input type="text" id="name" placeholder="" name="name"/>
+        </div>
+        <div class="pure-control-group">
+            <label for="aligned-age">Patient Age *</label>
+            <input type="number" id="age" placeholder="" name="age"/>
+        </div>
+        <div class="pure-control-group">
+            <label for="aligned-admission">Date of Admission *</label>
+            <input type="date" id="admission" placeholder="" name="dateOfAdmission" />
+        </div>
+        <div><label style="padding-left: 80px" for="aligned-states">Type of bed *</label>
+         <select name="typeOfBed">
+         	<option value="">Select type..</option>
+         	<option value="type-1">type-1</option>
+         	<option value="type-2">type-2</option>
+         	<option value="type-3">type-3</option>
+         </select>
+     </div><br>
+              <div class="pure-control-group">
+            <label for="aligned-address">address *</label>
+            <input type="textarea" name="address"></input>
+            </div>
+
+          <div class="pure-control-group">
+			<label for="State">State</label>
+			<select id="select1"  name="state" onchange="populate(this.id,'select2')">
+		 <option value="">Select</option>
+                            <option value="Delhi">Delhi</option>
+                            <option value="Haryana">Haryana</option>
+                            <option value="Himachal Pradesh">Himachal Pradesh</option>
+                            <option value="Jammu and Kashmir">Jammu and Kashmir</option>
+                            <option value="Ladakh">Ladakh</option>
+                            <option value="Punjab">Punjab</option>
+                            <option value="Rajasthan">Rajasthan</option>
+                            <option value="Uttarakhand">Uttarakhand</option>
+                            <option value="Uttarpradesh">Uttar Pradesh</option>
+         		</select>
+		</div>     <div style="padding-left: 150px">
+		            <td>City</td>
+                    <td>
+                        <select  id="select2"  name="city">
+                            <option value="">Select</option>
+                        </select>
+                    </td>
+                </div>
+             <div class="pure-controls">
+            
+            <button type="submit" class="pure-button pure-button-primary">Submit</button>
+        </div>
+    </fieldset>
+</form>
+</body>
+</html>
+
+    </div>
+{% endblock %}
